# optimised IList / NonEmptyList constructors