I've been doing some benchmarking and the varargs constructors (everywhere in the codebase) are a problem. It causes a `Seq` to be created, and then a further `List`.

There is not much we can do in 7.2, but for 7.3+ we should

- have explicit arity constructors
- (optionally) have a varargs after that

but there is arguably limited value in a varargs constructor anyway because of the performance hit.

Here I have added some binary compatible (and source equivalent) methods that should speed up a lot of codepaths. But I can't catch zero, 1 and 2 arguments because of source incompatible conflicts.