@fommil 2018-05-18
@hrhino it creates a list when it does `: _*`, unless we have a sneaky `Seq` constructor somewhere that I couldn't find.