@hrhino 2018-05-18
huh. I thought it would just pass through the `WrappedArray` from the outer varargs. sigh