@fommil 2018-05-18
before (smaller is better)

```
[info] Benchmark                 Mode  Cnt    Score    Error  Units
[info] MapBench.ilist100         avgt    5    0.868 ±  0.020  us/op
[info] MapBench.ilist1000        avgt    5    9.019 ±  0.123  us/op
[info] MapBench.ilist10000       avgt    5   88.071 ±  9.789  us/op
[info] MapBench.ilistClass100    avgt    5    0.892 ±  0.035  us/op
[info] MapBench.ilistClass1000   avgt    5    9.307 ±  0.225  us/op
[info] MapBench.ilistClass10000  avgt    5  101.403 ± 10.516  us/op
```

after

```
[info] Benchmark                 Mode  Cnt   Score    Error  Units
[info] MapBench.ilist100         avgt    5   0.782 ±  0.091  us/op
[info] MapBench.ilist1000        avgt    5   7.885 ±  0.155  us/op
[info] MapBench.ilist10000       avgt    5  74.436 ± 11.774  us/op
[info] MapBench.ilistClass100    avgt    5   0.751 ±  0.005  us/op
[info] MapBench.ilistClass1000   avgt    5   7.725 ±  0.237  us/op
[info] MapBench.ilistClass10000  avgt    5  83.969 ±  1.246  us/op
```