@jdegoes 2018-05-18
@fommil I agree vararg constructors are a bad idea from performance and should be avoided in the future. Nice work in optimizing this. 😄